// src/app/pages/vehiculos/components/vehiculos-table/vehiculos.interface.ts
export interface VehiculosInterface {
  idvehiculo?: number;
  propietario: number;
  anio: number;
  marca: string;
  kilometraje: number;
  modelo: string;
  serie: string;
  serieMotor:string;
  placa:string;
  status: string;
  poliza:string;
  polizaTipo:string;
  condActual: string;
  condInicial:string;
  color:string;
}